# Support task resource tracking in OpenSearch