--- a/server/src/internalClusterTest/java/org/opensearch/action/admin/cluster/node/tasks/TasksIT.java
+++ b/server/src/internalClusterTest/java/org/opensearch/action/admin/cluster/node/tasks/TasksIT.java
@@ -470,6 +470,9 @@ public class TasksIT extends OpenSearchIntegTestCase {
 
                     @Override
                     public void waitForTaskCompletion(Task task) {}
+
+                    @Override
+                    public void taskExecutionStarted(Task task, Boolean closeableInvoked) {}
                 });
             }
             // Need to run the task in a separate thread because node client's .execute() is blocked by our task listener
@@ -650,6 +653,9 @@ public class TasksIT extends OpenSearchIntegTestCase {
                     public void waitForTaskCompletion(Task task) {
                         waitForWaitingToStart.countDown();
                     }
+
+                    @Override
+                    public void taskExecutionStarted(Task task, Boolean closeableInvoked) {}
 
                     @Override
                     public void onTaskRegistered(Task task) {}