--- a/server/src/main/java/org/opensearch/action/search/SearchShardTask.java
+++ b/server/src/main/java/org/opensearch/action/search/SearchShardTask.java
@@ -52,6 +52,11 @@ public class SearchShardTask extends CancellableTask {
     }
 
     @Override
+    public boolean supportsResourceTracking() {
+        return true;
+    }
+
+    @Override
     public boolean shouldCancelChildrenOnCancellation() {
         return false;
     }