--- a/server/src/main/java/org/opensearch/tasks/Task.java
+++ b/server/src/main/java/org/opensearch/tasks/Task.java
@@ -32,8 +32,6 @@
 
 package org.opensearch.tasks;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.opensearch.action.ActionResponse;
 import org.opensearch.cluster.node.DiscoveryNode;
 import org.opensearch.common.io.stream.NamedWriteable;
@@ -54,8 +52,6 @@ import java.util.concurrent.ConcurrentHashMap;
  * @opensearch.internal
  */
 public class Task {
-
-    private static final Logger logger = LogManager.getLogger(Task.class);
 
     /**
      * The request header to mark tasks with specific ids
@@ -291,7 +287,7 @@ public class Task {
                 );
             }
         }
-        threadResourceInfoList.add(new ThreadResourceInfo(statsType, resourceUsageMetrics));
+        threadResourceInfoList.add(new ThreadResourceInfo(threadId, statsType, resourceUsageMetrics));
     }
 
     /**
@@ -336,6 +332,17 @@ public class Task {
             }
         }
         throw new IllegalStateException("cannot update final values if active thread resource entry is not present");
+    }
+
+    /**
+     * Individual tasks can override this if they want to support task resource tracking. We just need to make sure that
+     * the ThreadPool on which the task runs on have runnable wrapper similar to
+     * {@link org.opensearch.common.util.concurrent.OpenSearchExecutors#newResizable}
+     *
+     * @return true if resource tracking is supported by the task
+     */
+    public boolean supportsResourceTracking() {
+        return false;
     }
 
     /**