### Description

Reopens changes from #2639 (reverted in #3046) to add a framework for task resource tracking.
Currently, SearchTask and SearchShardTask support resource tracking but it can be extended to any other task in the future.

Changes since #2639:
* Replaced the usage of AutoQueueAdjustingExecutorBuilder with ResizableExecutorBuilder
* Fixed a race-condition when Task is unregistered before its threads are stopped
* Resolved merge conflicts
* Fixed broken tests

Signed-off-by: Ketan Verma <ketan9495@gmail.com>
 
### Issues Resolved
#1179
 
### Check List
- [x] New functionality includes testing.
  - [x] All tests pass
- [x] New functionality has been documented.
  - [x] New functionality has javadoc added
- [x] Commits are signed per the DCO using --signoff 

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.
For more information on following Developer Certificate of Origin and signing off your commits, please check [here](https://github.com/opensearch-project/OpenSearch/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).