@ketanv3 2022-07-23
> Thanks @ketanv3 for the changes. What is the additional delay the await mechanism might introduce. We might need to run benchmarks for this change

Yes, I'm working towards running performance benchmarks.