@ketanv3 2022-07-31
Comparing the accuracy of both approaches – (1) no waiting for task's threads to complete, and (2) using a callback to keep track of active threads – before marking the task as unregistered.

An existing integration test was re-used to perform large number of search requests with predictable CPU/memory usage. Measurements were taken for:
* Number of times thread usage was reported before task tracking was stopped (thread usage accounted).
* Number of times thread usage was reported after task tracking was stopped (thread usage lost).

Both tests executed 500 search queries and reported resource usages for ~9020 tasks.
| | approach 1 | approach 2 |
|-|------------| ------------|
| thread usage accounted | 2641 | 4392 |
| thread usage lost | 3721 | 237 |
| tasks completed | 9020 | 9021 |

Based on these results, approach (2) has been used for the implementation as it gives better accuracy.