@ketanv3 2022-08-01
## Benchmark results

* Used c5.2xlarge EC2 instance-type
* Used `./gradlew localDistro` to generate a distribution from arbitrary commits
* Used the default configs to launch a cluster
* Used `nyc_taxis` workload

Baseline commit: 740f75d2051
Contender commit: a09a60acbbb

```
opensearch-benchmark compare --baseline 740f75d2051 --contender a09a60acbbb

   ____                  _____                      __       ____                  __                         __
  / __ \____  ___  ____ / ___/___  ____ ___________/ /_     / __ )___  ____  _____/ /_  ____ ___  ____ ______/ /__
 / / / / __ \/ _ \/ __ \\__ \/ _ \/ __ `/ ___/ ___/ __ \   / __  / _ \/ __ \/ ___/ __ \/ __ `__ \/ __ `/ ___/ //_/
/ /_/ / /_/ /  __/ / / /__/ /  __/ /_/ / /  / /__/ / / /  / /_/ /  __/ / / / /__/ / / / / / / / / /_/ / /  / ,<
\____/ .___/\___/_/ /_/____/\___/\__,_/_/   \___/_/ /_/  /_____/\___/_/ /_/\___/_/ /_/_/ /_/ /_/\__,_/_/  /_/|_|
    /_/


Comparing baseline
  TestExecution ID: 740f75d2051
  TestExecution timestamp: 2022-07-31 13:29:54
  TestProcedure: append-no-conflicts
  ProvisionConfigInstance: external

with contender
  TestExecution ID: a09a60acbbb
  TestExecution timestamp: 2022-08-01 09:25:35
  TestProcedure: append-no-conflicts
  ProvisionConfigInstance: external

------------------------------------------------------
    _______             __   _____
   / ____(_)___  ____ _/ /  / ___/_________  ________
  / /_  / / __ \/ __ `/ /   \__ \/ ___/ __ \/ ___/ _ \
 / __/ / / / / / /_/ / /   ___/ / /__/ /_/ / /  /  __/
/_/   /_/_/ /_/\__,_/_/   /____/\___/\____/_/   \___/
------------------------------------------------------

|                                                        Metric |                     Task |    Baseline |   Contender |     Diff |   Unit |
|--------------------------------------------------------------:|-------------------------:|------------:|------------:|---------:|-------:|
|                    Cumulative indexing time of primary shards |                          |       126.4 |     126.926 |   0.5262 |    min |
|             Min cumulative indexing time across primary shard |                          |       126.4 |     126.926 |   0.5262 |    min |
|          Median cumulative indexing time across primary shard |                          |       126.4 |     126.926 |   0.5262 |    min |
|             Max cumulative indexing time across primary shard |                          |       126.4 |     126.926 |   0.5262 |    min |
|           Cumulative indexing throttle time of primary shards |                          |           0 |           0 |        0 |    min |
|    Min cumulative indexing throttle time across primary shard |                          |           0 |           0 |        0 |    min |
| Median cumulative indexing throttle time across primary shard |                          |           0 |           0 |        0 |    min |
|    Max cumulative indexing throttle time across primary shard |                          |           0 |           0 |        0 |    min |
|                       Cumulative merge time of primary shards |                          |     52.5814 |      54.486 |  1.90467 |    min |
|                      Cumulative merge count of primary shards |                          |         208 |         212 |        4 |        |
|                Min cumulative merge time across primary shard |                          |     52.5814 |      54.486 |  1.90467 |    min |
|             Median cumulative merge time across primary shard |                          |     52.5814 |      54.486 |  1.90467 |    min |
|                Max cumulative merge time across primary shard |                          |     52.5814 |      54.486 |  1.90467 |    min |
|              Cumulative merge throttle time of primary shards |                          |     1.37742 |     1.66843 |  0.29102 |    min |
|       Min cumulative merge throttle time across primary shard |                          |     1.37742 |     1.66843 |  0.29102 |    min |
|    Median cumulative merge throttle time across primary shard |                          |     1.37742 |     1.66843 |  0.29102 |    min |
|       Max cumulative merge throttle time across primary shard |                          |     1.37742 |     1.66843 |  0.29102 |    min |
|                     Cumulative refresh time of primary shards |                          |    0.690283 |    0.595333 | -0.09495 |    min |
|                    Cumulative refresh count of primary shards |                          |          76 |          80 |        4 |        |
|              Min cumulative refresh time across primary shard |                          |    0.690283 |    0.595333 | -0.09495 |    min |
|           Median cumulative refresh time across primary shard |                          |    0.690283 |    0.595333 | -0.09495 |    min |
|              Max cumulative refresh time across primary shard |                          |    0.690283 |    0.595333 | -0.09495 |    min |
|                       Cumulative flush time of primary shards |                          |     1.51388 |     1.44558 |  -0.0683 |    min |
|                      Cumulative flush count of primary shards |                          |          31 |          33 |        2 |        |
|                Min cumulative flush time across primary shard |                          |     1.51388 |     1.44558 |  -0.0683 |    min |
|             Median cumulative flush time across primary shard |                          |     1.51388 |     1.44558 |  -0.0683 |    min |
|                Max cumulative flush time across primary shard |                          |     1.51388 |     1.44558 |  -0.0683 |    min |
|                                       Total Young Gen GC time |                          |      64.188 |      66.636 |    2.448 |      s |
|                                      Total Young Gen GC count |                          |       17754 |       18124 |      370 |        |
|                                         Total Old Gen GC time |                          |           0 |           0 |        0 |      s |
|                                        Total Old Gen GC count |                          |           0 |           0 |        0 |        |
|                                                    Store size |                          |     24.3704 |     24.3655 | -0.00482 |     GB |
|                                                 Translog size |                          | 5.12227e-08 | 5.12227e-08 |        0 |     GB |
|                                        Heap used for segments |                          |           0 |           0 |        0 |     MB |
|                                      Heap used for doc values |                          |           0 |           0 |        0 |     MB |
|                                           Heap used for terms |                          |           0 |           0 |        0 |     MB |
|                                           Heap used for norms |                          |           0 |           0 |        0 |     MB |
|                                          Heap used for points |                          |           0 |           0 |        0 |     MB |
|                                   Heap used for stored fields |                          |           0 |           0 |        0 |     MB |
|                                                 Segment count |                          |          30 |          30 |        0 |        |
|                                                Min Throughput |                    index |      139816 |      140101 |  284.462 | docs/s |
|                                               Mean Throughput |                    index |      140748 |      141834 |  1086.13 | docs/s |
|                                             Median Throughput |                    index |      140712 |      141860 |  1147.85 | docs/s |
|                                                Max Throughput |                    index |      141949 |      143104 |   1155.2 | docs/s |
|                                       50th percentile latency |                    index |      459.04 |     461.438 |  2.39825 |     ms |
|                                       90th percentile latency |                    index |     641.768 |     667.389 |  25.6212 |     ms |
|                                       99th percentile latency |                    index |      1401.7 |     1471.98 |  70.2789 |     ms |
|                                     99.9th percentile latency |                    index |     2154.13 |     2254.37 |  100.243 |     ms |
|                                    99.99th percentile latency |                    index |     2801.73 |      2947.4 |   145.67 |     ms |
|                                      100th percentile latency |                    index |     2950.47 |     3561.21 |  610.743 |     ms |
|                                  50th percentile service time |                    index |      459.04 |     461.438 |  2.39825 |     ms |
|                                  90th percentile service time |                    index |     641.768 |     667.389 |  25.6212 |     ms |
|                                  99th percentile service time |                    index |      1401.7 |     1471.98 |  70.2789 |     ms |
|                                99.9th percentile service time |                    index |     2154.13 |     2254.37 |  100.243 |     ms |
|                               99.99th percentile service time |                    index |     2801.73 |      2947.4 |   145.67 |     ms |
|                                 100th percentile service time |                    index |     2950.47 |     3561.21 |  610.743 |     ms |
|                                                    error rate |                    index |           0 |           0 |        0 |      % |
|                                                Min Throughput | wait-until-merges-finish |  0.00545798 |  0.00717392 |  0.00172 |  ops/s |
|                                               Mean Throughput | wait-until-merges-finish |  0.00545798 |  0.00717392 |  0.00172 |  ops/s |
|                                             Median Throughput | wait-until-merges-finish |  0.00545798 |  0.00717392 |  0.00172 |  ops/s |
|                                                Max Throughput | wait-until-merges-finish |  0.00545798 |  0.00717392 |  0.00172 |  ops/s |
|                                      100th percentile latency | wait-until-merges-finish |      183218 |      139394 | -43824.2 |     ms |
|                                 100th percentile service time | wait-until-merges-finish |      183218 |      139394 | -43824.2 |     ms |
|                                                    error rate | wait-until-merges-finish |           0 |           0 |        0 |      % |
|                                                Min Throughput |                  default |     3.01589 |     3.01582 |   -7e-05 |  ops/s |
|                                               Mean Throughput |                  default |     3.02592 |      3.0258 | -0.00011 |  ops/s |
|                                             Median Throughput |                  default |     3.02365 |     3.02349 | -0.00016 |  ops/s |
|                                                Max Throughput |                  default |     3.04569 |     3.04548 | -0.00022 |  ops/s |
|                                       50th percentile latency |                  default |      5.7814 |     5.91026 |  0.12886 |     ms |
|                                       90th percentile latency |                  default |     6.43334 |     6.83153 |  0.39818 |     ms |
|                                       99th percentile latency |                  default |     8.34059 |     10.2823 |  1.94167 |     ms |
|                                      100th percentile latency |                  default |      9.3404 |     11.1879 |  1.84747 |     ms |
|                                  50th percentile service time |                  default |     3.23102 |     3.30837 |  0.07735 |     ms |
|                                  90th percentile service time |                  default |     3.64748 |     3.97348 |  0.32599 |     ms |
|                                  99th percentile service time |                  default |     5.69504 |     7.69124 |   1.9962 |     ms |
|                                 100th percentile service time |                  default |      6.9198 |     8.15025 |  1.23045 |     ms |
|                                                    error rate |                  default |           0 |           0 |        0 |      % |
|                                                Min Throughput |                    range |    0.703708 |    0.703334 | -0.00037 |  ops/s |
|                                               Mean Throughput |                    range |    0.706099 |    0.705482 | -0.00062 |  ops/s |
|                                             Median Throughput |                    range |    0.705548 |    0.704986 | -0.00056 |  ops/s |
|                                                Max Throughput |                    range |    0.711012 |    0.709893 | -0.00112 |  ops/s |
|                                       50th percentile latency |                    range |     230.942 |     228.752 | -2.18983 |     ms |
|                                       90th percentile latency |                    range |     232.469 |      232.27 | -0.19822 |     ms |
|                                       99th percentile latency |                    range |     281.604 |      268.33 | -13.2736 |     ms |
|                                      100th percentile latency |                    range |     282.024 |     274.873 | -7.15022 |     ms |
|                                  50th percentile service time |                    range |     224.228 |     221.927 | -2.30112 |     ms |
|                                  90th percentile service time |                    range |     225.436 |     225.098 | -0.33821 |     ms |
|                                  99th percentile service time |                    range |     274.534 |     261.367 | -13.1669 |     ms |
|                                 100th percentile service time |                    range |     274.651 |     268.094 |  -6.5567 |     ms |
|                                                    error rate |                    range |           0 |           0 |        0 |      % |
|                                                Min Throughput |      distance_amount_agg |     2.01208 |     2.01214 |    6e-05 |  ops/s |
|                                               Mean Throughput |      distance_amount_agg |     2.01986 |     2.01997 |  0.00011 |  ops/s |
|                                             Median Throughput |      distance_amount_agg |     2.01805 |     2.01815 |   0.0001 |  ops/s |
|                                                Max Throughput |      distance_amount_agg |     2.03565 |      2.0359 |  0.00025 |  ops/s |
|                                       50th percentile latency |      distance_amount_agg |     5.11124 |     5.33593 |  0.22468 |     ms |
|                                       90th percentile latency |      distance_amount_agg |     5.49016 |      5.5372 |  0.04703 |     ms |
|                                       99th percentile latency |      distance_amount_agg |     5.91635 |     5.86624 | -0.05011 |     ms |
|                                      100th percentile latency |      distance_amount_agg |     5.94142 |     6.09461 |  0.15319 |     ms |
|                                  50th percentile service time |      distance_amount_agg |     1.83493 |     1.90533 |   0.0704 |     ms |
|                                  90th percentile service time |      distance_amount_agg |     2.09502 |     2.09756 |  0.00254 |     ms |
|                                  99th percentile service time |      distance_amount_agg |     2.26754 |     2.35835 |  0.09081 |     ms |
|                                 100th percentile service time |      distance_amount_agg |     2.44594 |     2.45766 |  0.01172 |     ms |
|                                                    error rate |      distance_amount_agg |           0 |           0 |        0 |      % |
|                                                Min Throughput |            autohisto_agg |     1.50055 |     1.50018 | -0.00037 |  ops/s |
|                                               Mean Throughput |            autohisto_agg |     1.50088 |     1.50029 | -0.00059 |  ops/s |
|                                             Median Throughput |            autohisto_agg |     1.50081 |     1.50027 | -0.00055 |  ops/s |
|                                                Max Throughput |            autohisto_agg |     1.50158 |      1.5005 | -0.00108 |  ops/s |
|                                       50th percentile latency |            autohisto_agg |     432.987 |     447.084 |  14.0965 |     ms |
|                                       90th percentile latency |            autohisto_agg |     440.025 |     454.874 |   14.849 |     ms |
|                                       99th percentile latency |            autohisto_agg |     445.834 |     462.858 |  17.0243 |     ms |
|                                      100th percentile latency |            autohisto_agg |     452.361 |     466.995 |  14.6342 |     ms |
|                                  50th percentile service time |            autohisto_agg |     430.369 |      445.16 |  14.7912 |     ms |
|                                  90th percentile service time |            autohisto_agg |     437.996 |     452.532 |  14.5358 |     ms |
|                                  99th percentile service time |            autohisto_agg |       443.3 |     460.938 |  17.6376 |     ms |
|                                 100th percentile service time |            autohisto_agg |     450.203 |     463.866 |  13.6633 |     ms |
|                                                    error rate |            autohisto_agg |           0 |           0 |        0 |      % |
|                                                Min Throughput |       date_histogram_agg |     1.50276 |     1.50321 |  0.00044 |  ops/s |
|                                               Mean Throughput |       date_histogram_agg |     1.50448 |     1.50523 |  0.00075 |  ops/s |
|                                             Median Throughput |       date_histogram_agg |     1.50409 |     1.50477 |  0.00068 |  ops/s |
|                                                Max Throughput |       date_histogram_agg |     1.50791 |     1.50925 |  0.00134 |  ops/s |
|                                       50th percentile latency |       date_histogram_agg |     460.007 |      446.69 | -13.3169 |     ms |
|                                       90th percentile latency |       date_histogram_agg |     469.084 |     455.775 | -13.3096 |     ms |
|                                       99th percentile latency |       date_histogram_agg |     478.717 |      467.45 | -11.2664 |     ms |
|                                      100th percentile latency |       date_histogram_agg |     491.838 |     468.051 | -23.7876 |     ms |
|                                  50th percentile service time |       date_histogram_agg |     457.659 |     444.775 | -12.8836 |     ms |
|                                  90th percentile service time |       date_histogram_agg |     466.225 |     453.912 | -12.3126 |     ms |
|                                  99th percentile service time |       date_histogram_agg |     475.816 |     465.039 |  -10.777 |     ms |
…